marketing_ai/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.0
pydantic[email]==2.5.2
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
celery==5.3.6
redis==5.0.1
qdrant-client>=1.7.0
stripe>=7.13.0
openai==1.3.5
anthropic==0.7.7
# airllm==0.0.1  # Assuming placeholder for airllm integration
beautifulsoup4==4.12.2
httpx==0.25.2
cryptography==41.0.7
pandas==2.1.3
numpy==1.26.2
